Spostata la logica di importazione all'interno di Management
