+++ Api/ImportBlockManagementInterface.php
@@ -1,0 +1,7 @@
+<?php
+namespace DNAFactory\DataSeeder\Api;
+
+interface ImportBlockManagementInterface 
+{
+    public function import($fileName = "blocks.php");
+}
