+++ Api/ImportConfigManagementInterface.php
@@ -1,0 +1,7 @@
+<?php
+namespace DNAFactory\DataSeeder\Api;
+
+interface ImportConfigManagementInterface
+{
+    public function import($fileName = "config", $environment = "production");
+}
