+++ Api/ImportEmailManagementInterface.php
@@ -1,0 +1,7 @@
+<?php
+namespace DNAFactory\DataSeeder\Api;
+
+interface ImportEmailManagementInterface 
+{
+    public function import($sqlFilename = "email_template.sql");
+}
