+++ Api/ImportPageManagementInterface.php
@@ -1,0 +1,7 @@
+<?php
+namespace DNAFactory\DataSeeder\Api;
+
+interface ImportPageManagementInterface 
+{
+    public function import($filename = "pages.php");
+}
