+++ Command/ImportPageCommand.php
@@ -1,11 +1,8 @@
 <?php
 namespace DNAFactory\DataSeeder\Command;
 
-use DNAFactory\DataSeeder\Asset\FetchAssetContent;
-use DNAFactory\DataSeeder\Asset\PageIntegrityResolver;
-use Magento\Cms\Api\PageRepositoryInterface;
-use Magento\Cms\Model\Page;
-use Magento\Cms\Model\PageFactory;
+use DNAFactory\DataSeeder\Api\ImportPageManagementInterface;
+
 use Magento\Framework\App\Config\Storage\WriterInterface;
 use Magento\Framework\Module\Dir\Reader;
 use Symfony\Component\Console\Command\Command;
@@ -15,62 +12,22 @@ use Symfony\Component\Console\Output\OutputInterface;
 
 class ImportPageCommand extends Command
 {
-    const PAGE_PATH = 'pages/';
-    const PAGE_CONTENT_PATH = self::PAGE_PATH . 'contents/';
-
     /**
-     * @var Reader
+     * @var ImportPageManagementInterface
      */
-    protected $directoryReader;
-    /**
-     * @var WriterInterface
-     */
-    protected $configWriter;
-    /**
-     * @var InputInterface
-     */
-    protected $input;
-    /**
-     * @var OutputInterface
-     */
-    protected $output;
-    /**
-     * @var FetchAssetContent
-     */
-    private $fetchAssetContent;
-    /**
-     * @var PageIntegrityResolver
-     */
-    private $integrityResolver;
-    /**
-     * @var PageFactory
-     */
-    private $pageFactory;
-    /**
-     * @var PageRepositoryInterface
-     */
-    private $pageRepository;
-
+    protected $importPageManagement;
+    
     /**
      * ImportBlockCommand constructor.
-     * @param PageFactory $pageFactory
-     * @param PageRepositoryInterface $pageRepository
-     * @param FetchAssetContent $fetchAssetContent
-     * @param PageIntegrityResolver $integrityResolver
+     * @param ImportPageManagementInterface $importPageManagement
      * @param string|null $name
      */
     public function __construct(
-        PageFactory $pageFactory,
-        PageRepositoryInterface $pageRepository,
-        FetchAssetContent $fetchAssetContent,
-        PageIntegrityResolver $integrityResolver,
+        ImportPageManagementInterface $importPageManagement,
         string $name = null
     ) {
         parent::__construct($name);
-        $this->fetchAssetContent = $fetchAssetContent;
-        $this->integrityResolver = $integrityResolver;
-        $this->pageFactory = $pageFactory;
-        $this->pageRepository = $pageRepository;
+        $this->importPageManagement = $importPageManagement;
     }
 
     protected function configure()
@@ -88,27 +45,7 @@ class ImportPageCommand extends Command
         $this->input = $input;
         $this->output = $output;
 
-        $pages = $this->getPages($input->getArgument('page-filename'));
-
-        foreach ($pages as $page) {
-            try {
-                $content = $this->fetchAssetContent->execute(self::PAGE_CONTENT_PATH . $page["file_name"]);
-
-                $this->integrityResolver->execute($page['page_data']['identifier']);
-                /** @var Page $pageModel */
-                $pageModel = $this->pageFactory->create();
-                $pageModel->setData($page['page_data']);
-                $pageModel->setContent($content);
-                $this->pageRepository->save($pageModel);
-            } catch (\Exception $exception) {
-                echo $exception->getMessage();
-            }
-        }
-    }
-
-    protected function getPages($filename)
-    {
-        $filePagesPath = $this->fetchAssetContent->getRootPath() . self::PAGE_PATH . $filename;
-        return include $filePagesPath;
+        $filename = $input->getArgument('page-filename');
+        $this->importPageManagement->import($filename);
     }
 }
