+++ composer.json
@@ -1,7 +1,7 @@
 {
   "name": "dnafactory/module-dataseeder",
   "type": "magento2-module",
-  "version": "1.0.1",
+  "version": "1.0.3",
   "description": "This module helps you to save and import some of magento 2 entities (such as Blocks, Pages, Configurations and Email).",
   "license": [
   ],
@@ -17,6 +17,11 @@
     {
       "name": "Antonio Vitale",
       "email": "antonio.vitale@dnafactory.it",
+      "role": "Developer"
+    },
+    {
+      "name": "Luigi Ambra",
+      "email": "luigi.ambra@dnafactory.it",
       "role": "Developer"
     }
